Separate the empty functions so code splitting can work properly.
