+++ projects/aem-angular-core-wcm-components/authoring/button/v1/src/public-api.ts
@@ -13,6 +13,8 @@
  ~ See the License for the specific language governing permissions and
  ~ limitations under the License.
  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
+import {ButtonV1Model, BreadCrumbV2IsEmptyFn} from "@adobe/aem-core-components-angular-base/core";
+export {ButtonV1Model, BreadCrumbV2IsEmptyFn};
 
 export * from "./button.v1.component";
 export * from "./button.v1.module";
