+++ projects/aem-angular-core-wcm-components/authoring/default/v1/src/public-api.ts
@@ -13,6 +13,9 @@
  ~ See the License for the specific language governing permissions and
  ~ limitations under the License.
  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
+import {DefaultV1Model, DefaultV1IsEmptyFn} from "@adobe/aem-core-components-angular-base/core";
+export {DefaultV1Model, DefaultV1IsEmptyFn}
 
 export * from "./default.v1.component";
 export * from "./default.v1.module";
+
