+++ projects/aem-angular-core-wcm-components/authoring/download/v1/src/public-api.ts
@@ -16,3 +16,6 @@
 
 export * from "./download.v1.component";
 export * from "./download.v1.module";
+
+import {DownloadV1IsEmptyFn, DownloadV1LabelsModel, DefaultV1Model} from "@adobe/aem-core-components-angular-base/core";
+export {DownloadV1IsEmptyFn, DownloadV1LabelsModel, DefaultV1Model}
