+++ projects/aem-angular-core-wcm-components/authoring/list/v2/src/list.v2.component.ts
@@ -17,30 +17,13 @@
 
 
 import {Component, Input} from "@angular/core";
-import {AbstractRoutedCoreComponent, RoutedCoreComponentModel} from "@adobe/aem-core-components-angular-base/core";
+import {
+    AbstractRoutedCoreComponent, ListV2IsEmptyFn, ListV2Item,
+    ListV2Model,
+    RoutedCoreComponentModel
+} from "@adobe/aem-core-components-angular-base/core";
 
 
-export interface ListV2Item extends RoutedCoreComponentModel{
-    url?:string
-    lastModified?:number
-    lastModifiedFormatted?:string
-    description?:string
-    path:string
-    title:string
-}
-
-
-export interface ListV2Model extends RoutedCoreComponentModel{
-    items:ListV2Item[]
-    dateFormatString: string
-    showDescription: boolean
-    showModificationDate: boolean
-    linkItems: boolean
-}
-
-export function ListV2IsEmptyFn(props:ListV2Model): boolean{
-    return props.items == null || props.items.length === 0;
-}
 
 
 @Component({
