+++ projects/aem-angular-core-wcm-components/authoring/list/v2/src/public-api.ts
@@ -16,3 +16,6 @@
 
 export * from "./list.v2.component";
 export * from "./list.v2.module";
+
+import {ListV2IsEmptyFn, ListV2Item, ListV2Model} from "@adobe/aem-core-components-angular-base/core";
+export {ListV2IsEmptyFn, ListV2Item, ListV2Model}
