+++ projects/aem-angular-core-wcm-components/authoring/teaser/v1/src/public-api.ts
@@ -16,3 +16,6 @@
 
 export * from "./teaser.v1.component";
 export * from "./teaser.v1.module";
+
+import {TeaserV1IsEmptyFn, TeaserV1Action, TeaserV1Model} from "@adobe/aem-core-components-angular-base/core";
+export {TeaserV1IsEmptyFn, TeaserV1Action, TeaserV1Model}
