+++ projects/aem-angular-core-wcm-components/authoring/text/v2/src/public-api.ts
@@ -16,3 +16,6 @@
 
 export * from "./text.v2.component";
 export * from "./text.v2.module";
+
+import {TextV2IsEmptyFn, TextV2Model} from "@adobe/aem-core-components-angular-base/core";
+export {TextV2IsEmptyFn, TextV2Model}
