+++ projects/aem-angular-core-wcm-components/authoring/text/v2/src/text.v2.component.ts
@@ -17,17 +17,8 @@
 
 
 import {Component, HostBinding, Input} from "@angular/core";
-import {AbstractCoreComponent} from "@adobe/aem-core-components-angular-base/core";
+import {AbstractCoreComponent, TextV2IsEmptyFn, TextV2Model} from "@adobe/aem-core-components-angular-base/core";
 
-
-export interface TextV2Model {
-    text?: string;
-    richText: boolean
-}
-
-export function TextV2IsEmptyFn(props:TextV2Model): boolean{
-    return props.text == null || props.text.trim().length === 0;
-}
 
 @Component({
     selector: 'core-text-v2',
