+++ projects/aem-angular-core-wcm-components/core/src/emptyfunctions/button.v1.isempty.ts
@@ -1,0 +1,12 @@
+import {RoutedCoreComponentModel} from "../model/RoutedCoreComponentModel";
+
+export interface ButtonV1Model extends RoutedCoreComponentModel{
+    text?: string;
+    link?: string;
+    icon?: string;
+    ariaLabel?: string;
+}
+
+export function ButtonV1IsEmptyFn(props:ButtonV1Model): boolean{
+    return props.text == null || props.text.trim().length === 0;
+}
