+++ projects/aem-angular-core-wcm-components/core/src/emptyfunctions/default.v1.isempty.ts
@@ -1,0 +1,8 @@
+
+export interface DefaultV1Model{
+    aHtml: string;
+}
+
+export function DefaultV1IsEmptyFn(props:DefaultV1Model): boolean{
+    return props.aHtml == null || props.aHtml.trim().length === 0;
+}
