+++ projects/aem-angular-core-wcm-components/core/src/emptyfunctions/image.v2.isempty.ts
@@ -1,0 +1,14 @@
+import {RoutedCoreComponentModel} from "../model/RoutedCoreComponentModel";
+
+export interface ImageV2Model extends RoutedCoreComponentModel{
+    src: string
+    alt: string
+    displayPopupTitle?: boolean
+    title?: string
+    link?: string
+}
+
+export function ImageV2IsEmptyFn(props:ImageV2Model) {
+    return (!props.src) || props.src.length === 0;
+}
+
