+++ projects/aem-angular-core-wcm-components/core/src/emptyfunctions/separator.v1.isempty.ts
@@ -1,0 +1,3 @@
+export function SeparatorV1IsEmptyFn(): boolean{
+    return false;
+}
