+++ projects/aem-angular-core-wcm-components/layout/breadcrumb/v2/src/public-api.ts
@@ -1,2 +1,5 @@
 export * from "./breadcrumb.v2.component";
 export * from "./breadcrumb.v2.module";
+
+import {BreadCrumbV2IsEmptyFn, BreadCrumbV2ItemModel, BreadCrumbV2Model} from "@adobe/aem-core-components-angular-base/core";
+export {BreadCrumbV2IsEmptyFn, BreadCrumbV2ItemModel, BreadCrumbV2Model}
