+++ projects/aem-angular-core-wcm-components/layout/language-navigation/v1/src/public-api.ts
@@ -1,3 +1,4 @@
 export * from "./language-navigation.v1.component";
 export * from "./language-navigation.v1.module";
 
+import {LanguageNavigationV1Item, LanguageNavigationV1Model} from "@adobe/aem-core-components-angular-base/core";
