+++ projects/aem-angular-core-wcm-components/layout/navigation/v1/src/navigation.v1.component.ts
@@ -19,29 +19,9 @@ import {
     AbstractRoutedCoreComponent,
     NAVIGATION_UTIL_SERVICE,
     NavigationItemModel,
-    NavigationUtilityService,
+    NavigationUtilityService, NavigationV1IsEmptyFn, NavigationV1Item, NavigationV1Model,
     RoutedCoreComponentModel
 } from "@adobe/aem-core-components-angular-base/core";
-
-export interface NavigationV1Item extends NavigationItemModel{
-    level: number,
-    active: boolean,
-    title: string,
-    url: string,
-    lastModified: number,
-    description?: string,
-    path: string,
-    children?: NavigationV1Item[]
-}
-
-export interface NavigationV1Model extends RoutedCoreComponentModel{
-    items:NavigationV1Item[]
-    accessibilityLabel?: string
-}
-
-export function NavigationV1IsEmptyFn(props:NavigationV1Model): boolean{
-    return props.items == null || props.items.length === 0;
-}
 
 
 @Component({
