+++ projects/aem-angular-core-wcm-components/layout/navigation/v1/src/public-api.ts
@@ -1,2 +1,5 @@
 export * from "./navigation.v1.component";
 export * from "./navigation.v1.module";
+
+import {NavigationV1IsEmptyFn, NavigationItemModel, NavigationV1Item} from "@adobe/aem-core-components-angular-base/core";
+export {NavigationV1IsEmptyFn, NavigationItemModel, NavigationV1Item}
